+++ resources/release-notes.md
@@ -1,3 +1,4 @@
 ## Software
 
 - (IMAP) Apply compression processing step before calibration corrections
+- (IMAP) Add ID `mag:time:ParseError` to error thrown for incorrect time formats in `mag.time.decodeTime`